+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - PHP 8.1 compatibility fixes
 - Upgraded to Predis v1.1.9
+- Added settings link to widget
 - Fixed wrong path to `diagnostics.php` file
 - Fixed chart overflow in settings tab